move components to lib folder, pascal case file names and fix imports
