+++ src/lib/components/EventRow.svelte
