+++ src/routes/+layout.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
   import { init } from "$lib/channel";
   import { onMount } from "svelte";
-  import Toasts from "../components/toasts.svelte";
+  import Toasts from "$lib/components/Toasts.svelte";
   import { topics } from "$lib/api";
-    import { seedMessages } from "$lib/api/data";
-    import { process } from "$lib/processor";
+  import { seedMessages } from "$lib/api/data";
+  import { process } from "$lib/processor";
 
   onMount(() => {
     // Hacky workaround to only call "init" once in a Svelte HMR life-cycle.
@@ -18,10 +18,10 @@
         for (const message of seedMessages) {
           await process(message);
         }
-        
+
         // After init subscribe to all calendars we know about.
         await topics.subscribeToAll();
-        
+
         // @TODO(sam): inform the backend of all authors we want to sync data with for each
         // calendar (add them to the topic map).
       });
