--- src/routes/app/events/[id]/+page.svelte
@@ -1,5 +1,0 @@
-<script>
-  import Header from "../../../../components/header.svelte";
-</script>
-
-<Header title="Event title" />
